Fixed uninstall to unmount all known mount points before nuking all
Now there is at least a chance to keep the files under /storage/*
if something goes wrong. The previous implementation did not unmount
those, and neither did it check if unmount succeeded thus creating
a situation where it could render the storage space empty (remember,
the script is being run as a root -> has all the power to remove
everything).