Added unmount command for system mount points
so that you could for example backup the system when it has just been
run. Or, to manually unmount all mount points before uninstalling.

Also, /data is not installed by default, only if --data or -d is
specified during install. If /data folder is present in the distro
installation then script will mount it.