Change to a new backup format and fix restoring backups
The previous implementation did not take in to account that the
backup was done with full path, and tried to restore at wrong
level. Also, because restoring at the root level poses a security
problem as well as problem with checking if the backup is for
the distro, or not, changed the directory level at which the
backups are made. Now, instead of creating the backup with full
path, the backup is now done relative to chroot-distro data path.

This way it is easy to verify that correct backup is being
processed and that backups can't affect the whole system.

While at it, added a check to ensure that the system mount points
are not mounted. This is because if they are mounted then their
contents will be also backed up. This poses a problem as when
the backup is being restored the backup is restored before system
mount points are in effect. This means that for example files
from external sdcard may also be backed up, and during restore
they will be saved to internal storage. Thus potentially rendering
the system unusable if the space is used up. Also, the restored
files will be shadowed by system mounts thus could lead to
confusion that why the files were not "restored", or that why
is my storage space full even though there does not seem to be
extra files.